All inventory moved to Inventory.java. Updated Verbs prepositions
