+++ Saves/playerSaves.json
@@ -1,1 +1,1 @@
-[{"saveFileId":0,"currentHP":8,"maxHP":8,"currentPosition":"Library","playerInventory":["Bread"],"currentStage":1,"weaponEquipped":"","weaponValue":0,"isUnderworld":false,"roomsVisited":[""]},{"saveFileId":1,"currentHP":4,"maxHP":8,"currentPosition":"Bedroom","playerInventory":["Bread"],"currentStage":1,"weaponEquipped":"","weaponValue":0,"isUnderworld":false,"roomsVisited":["Library","Hallway 3","Bedroom"]},{"saveFileId":2,"currentHP":2,"maxHP":8,"currentPosition":"Hallway3","playerInventory":["Bread"],"currentStage":2,"weaponEquipped":"","weaponValue":0,"isUnderworld":false,"roomsVisited":["Library","Bedroom"]},{"saveFileId":3,"currentHP":1,"maxHP":8,"currentPosition":"Library","playerInventory":["Bread"],"currentStage":1,"weaponEquipped":"","weaponValue":0,"isUnderworld":false,"roomsVisited":["Library"]},{"saveFileId":4,"currentHP":1,"maxHP":8,"currentPosition":"Library","playerInventory":["Bread"],"currentStage":1,"weaponEquipped":"","weaponValue":0,"isUnderworld":false,"roomsVisited":["Library"]},{"saveFileId":5,"currentHP":1,"maxHP":8,"currentPosition":"Library","playerInventory":["Bread"],"currentStage":1,"weaponEquipped":"","weaponValue":0,"isUnderworld":false,"roomsVisited":["Library"]},{"saveFileId":6,"currentHP":1,"maxHP":8,"currentPosition":"Library","playerInventory":["Bread"],"currentStage":1,"weaponEquipped":"","weaponValue":0,"isUnderworld":false,"roomsVisited":["Library"]},{"saveFileId":7,"currentHP":1,"maxHP":8,"currentPosition":"Library","playerInventory":["Bread"],"currentStage":1,"weaponEquipped":"","weaponValue":0,"isUnderworld":false,"roomsVisited":["Library"]},{"saveFileId":8,"currentHP":1,"maxHP":8,"currentPosition":"Library","playerInventory":["Bread"],"currentStage":1,"weaponEquipped":"","weaponValue":0,"isUnderworld":false,"roomsVisited":["Library"]},{"saveFileId":9,"currentHP":1,"maxHP":8,"currentPosition":"Library","playerInventory":["Bread"],"currentStage":1,"weaponEquipped":"","weaponValue":0,"isUnderworld":false,"roomsVisited":["Library"]},{"saveFileId":10,"currentHP":1,"maxHP":8,"currentPosition":"Library","playerInventory":["Bread"],"currentStage":1,"weaponEquipped":"","weaponValue":0,"isUnderworld":false,"roomsVisited":["Library"]}]
+[{"saveFileId":0,"currentHP":4,"maxHP":8,"currentPosition":"Hallway 2","playerInventory":["Journal"],"currentStage":2,"weaponEquipped":"","weaponValue":0,"isUnderworld":false,"roomsVisited":["Library","Hallway 3","Hallway 2"]},{"saveFileId":1,"currentHP":4,"maxHP":8,"currentPosition":"Bedroom","playerInventory":["Bread"],"currentStage":1,"weaponEquipped":"","weaponValue":0,"isUnderworld":false,"roomsVisited":["Library","Hallway 3","Bedroom"]},{"saveFileId":2,"currentHP":2,"maxHP":8,"currentPosition":"Hallway3","playerInventory":["Bread"],"currentStage":2,"weaponEquipped":"","weaponValue":0,"isUnderworld":false,"roomsVisited":["Library","Bedroom"]},{"saveFileId":3,"currentHP":1,"maxHP":8,"currentPosition":"Library","playerInventory":["Bread"],"currentStage":1,"weaponEquipped":"","weaponValue":0,"isUnderworld":false,"roomsVisited":["Library"]},{"saveFileId":4,"currentHP":1,"maxHP":8,"currentPosition":"Library","playerInventory":["Bread"],"currentStage":1,"weaponEquipped":"","weaponValue":0,"isUnderworld":false,"roomsVisited":["Library"]},{"saveFileId":5,"currentHP":1,"maxHP":8,"currentPosition":"Library","playerInventory":["Bread"],"currentStage":1,"weaponEquipped":"","weaponValue":0,"isUnderworld":false,"roomsVisited":["Library"]},{"saveFileId":6,"currentHP":1,"maxHP":8,"currentPosition":"Library","playerInventory":["Bread"],"currentStage":1,"weaponEquipped":"","weaponValue":0,"isUnderworld":false,"roomsVisited":["Library"]},{"saveFileId":7,"currentHP":1,"maxHP":8,"currentPosition":"Library","playerInventory":["Bread"],"currentStage":1,"weaponEquipped":"","weaponValue":0,"isUnderworld":false,"roomsVisited":["Library"]},{"saveFileId":8,"currentHP":1,"maxHP":8,"currentPosition":"Library","playerInventory":["Bread"],"currentStage":1,"weaponEquipped":"","weaponValue":0,"isUnderworld":false,"roomsVisited":["Library"]},{"saveFileId":9,"currentHP":1,"maxHP":8,"currentPosition":"Library","playerInventory":["Bread"],"currentStage":1,"weaponEquipped":"","weaponValue":0,"isUnderworld":false,"roomsVisited":["Library"]},{"saveFileId":10,"currentHP":1,"maxHP":8,"currentPosition":"Library","playerInventory":["Bread"],"currentStage":1,"weaponEquipped":"","weaponValue":0,"isUnderworld":false,"roomsVisited":["Library"]}]
